Trying CasperJS
Hit a roadblock with continuously getting 500's when using node, not sure
what the issue is but I decided I'll try using a headless browser instead.
Seamless doesn't do API keys anymore so I'll have to resort to virtually
clicking through the website.

The app will have to be run in two parts: 1 being the node.js server that
receives the order from Merakians and updates the Firebase entry accordingly,
and one is this casperJS bot that can retrieve data from Firebase and make
the orders accordingly.