Added ability to fetch from Heroku Alfred
The Google.com start is in there because I need to send a header containing
the token with the secret to Heroku Alfred. However, I can't find a way to send
headers through casper.start, and .start is needed to begin the chain. This is
such a dumb hack, I hope I can find a way to do this properly later.