Bug Fix: Exit if no orders
Previously, Casper would simply wait for timeout if there were no
orders. This caused a bug with the completion callback because it
would always call the unsuccessful ping in this case. So Casper
should exit earlier.

One interesting thing is that it took an hour and 10 minutes for it
to timeout and post to Alfred. I don't know why it took this long,
as it should only be 10 minutes.

Looking at the logs, it seems like the timeouts were building on top
of each other, with each waitFor not executing until the previous one
had finished waiting and timed out.