Changed timeout
It turns out that the waits are not done all at once, they are done
sequentially, so with 20 minutes timeout, each step could be waiting
for that long. With the addition of the error notification to Alfred,
we don't want to wait this long if something happens.

Oddly, the casper.on methods are not completely asynchronous, meaning
step 6 waits for step 5 to finish before starting its timeout. If this
were not the case, the addition of items, which almost always takes longer
than 20 seconds, would cause step 6 to timeout.

I believe this is because the extra .waits get chained when the item
addition starts, as removing the 'items.added' trigger times casper
out after all the items have been added.