Significantly speed up ordering process
I used to have arbitrary wait periods, this was safe but
unnecessarily caused Casper to take forever when ordering.
I've changed this to waitForSelectors, as most of these
actions actually effect an HTML change. I need to keep an
eye on Casper the next week or so to make sure nothing goes
wrong.