Possibly fix unsuccessful post to Alfie
For some reason, whenever an error occurs, Casper doesn't properly
post an unsuccessful order message to Alfie, it usually ends up being
invoking the 'about:blank' navigation bug. I'm not sure why that happens,
as it does try to "navigate" to Alfie first, but I'm going to try to
see if this delay alleviates that issue.