Change delivery time to ASAP
Linda is trying out different settings for the Seamless corporate
account. She has temporarily changed it so that orders can only be
placed (as in clicking the order button) after 5:15. We were setting
the delivery time as 5:15 because Little Delhi wasn't open before 5.
Now that we have to change the time at which Casper is triggered to
5:15, we can remove the time setting because it will default to
ASAP. This commit may be reverted and changed to a different time
in the future depending on what Linda decides to do.